* Replaced "ordering" filed of schedules by "starting_time" that records at what time a task should start * Added starting time filed in Task class * Added all embedded files in the list of files to pack * Updated xml schedule used as a correct reference in unit tests

+++ src/Architecture.cpp
@@ -58,7 +58,7 @@ namespace pelib
 	{
 		Algebra record;
 
-		Scalar<int> scalar_p("p", coreNumber);
+		Scalar<float> scalar_p("p", coreNumber);
 		Set<float> set_F("F", frequencies);
 
 		record.insert(&scalar_p);
